Minor name changes: Fother -> Parent (thought I fixed this at one point, but I must have missed some).
Dot and TripleDot are now properly called ellipses.

Also, some more preliminary work on reworking the entire class heirarchy from the ground up.  My initial idea is to have a Widget class that handles issues of sizing and position that inherits from an EventSource class that generalizes the spawning of events to event listeners.  The Window class would inherit from the Widget class, and extend into overlay and things of that nature.

This represents a rather large refactoring event, so I'll proceed in small atomic steps.