Added derived position
Changed coord to offset_derived.
Made it so ResourceLayouts can have position_derivied as a position
type.
Moved the derived calculation into the CoordConverter.
Made it so it is still based on width and height, but you can use
negative values for width and height, which then derives from the size
passed in.