added cmake script to build with SDL2
This patch only affects compilation with OpenGL as renderer.  For WIN32
SDL2 and SDL2_image are included in Dependencies folder.  for LINUX/MAC,
the packages need to be installed first.  Code wont compile yet in
MAC/LINUX since the #include in Base need to be modified.
Also, the four editors (font, layout, skin, image) are broken now, as
they use windows APIs.